+++ conf/urls.py
@@ -39,7 +39,17 @@ testapi_urls = [
     url(
         r'^submissions-by-email/(?P<email_address>.*)/$',
         testapi.views.SubmissionsTestAPIView.as_view(),
-        name='submissions-by-email',
+        name='submissions_by_email',
+    ),
+    url(
+        r'^test-senders/$',
+        testapi.views.SendersTestAPIView.as_view(),
+        name='delete_test_senders',
+    ),
+    url(
+        r'^test-submissions/$',
+        testapi.views.SubmissionsTestAPIView.as_view(),
+        name='delete_test_submissions',
     ),
 ]