+++ submission/admin.py
@@ -11,6 +11,32 @@ import core.helpers
 from submission import constants, models, tasks
 
 
+class FormUrlFilter(SimpleListFilter):
+    title = 'form URL'
+    parameter_name = 'form_url'
+    common_urls = [
+        '/contact/office-finder/*/',
+        '/international/trade/suppliers/*/contact/',
+        '/international/investment-support-directory/*/contact/',
+        '/suppliers/*/contact/',
+        '/investment-support-directory/*/contact/',
+    ]
+
+    def lookups(self, request, model_admin):
+        queryset = models.Submission.objects.all()
+        for url in self.common_urls:
+            escaped = url.replace('/', r'\/').replace('*', '.+')
+            queryset = queryset.exclude(form_url__iregex=f'{escaped}.*')
+        lookups = list(queryset.values_list('form_url', flat=True))
+        return [(item, item) for item in sorted(lookups + self.common_urls)]
+
+    def queryset(self, request, queryset):
+        value = self.value()
+        if value:
+            queryset = queryset.filter(form_url=value)
+        return queryset
+
+
 class ActionFilter(SimpleListFilter):
     title = 'action'
     parameter_name = 'action_name'
@@ -53,7 +79,7 @@ class SubmissionAdmin(core.helpers.DownloadCSVMixin, admin.ModelAdmin):
     list_filter = (
         'client',
         ActionFilter,
-        'form_url',
+        FormUrlFilter,
         'created',
         'is_sent',
         'sender',